Sai Sneha Yeldi/requirements.txt
streamlit
ollama
numpy
pillow
easyocr
torch
opencv-python
